Add files via upload
Sitecore-assessment for below scenarios

Exercise 1: Basic Azure knowledge
Create an environment with 2 machines.
The machines need to be able to communicate with each other.
Create a script which outputs basic metrics of the machines.
● CPU usage
● Network usage
● Disk usage
Please deliver:
● An automation script creating the environment
● A 2nd script outputting the metrics

Exercise 2: Basic Linux service setup
Create a small linux machine (ubuntu) to perform the following tasks:
● Get ElasticSearch running in a docker container
● Check the health of ElasticSearch from the command line
Please deliver:
● A bash script we can execute on an empty Linux (ubuntu) machine
We will evaluate the output of the batch script with regards to the requirements stated.
Exercise 3: Linux problem solving
Please connect with a SSH client to the following machine as mentioned in the intro using you SSH
key you provided to us.
We have an issue with Redis on this machine. It's not starting. Please diagnose and solve the issue.
Please deliver:
● A short description of your troubleshooting process.
● Tell us what's wrong
● Explain what you did to fix it
We will validate the machine and check if redis is running now.
Exercise 4: Windows problem solving
Please connect with an rdp client to the following machine as mentioned in the intro.
We have an issue with the web application installed. When browsing to http://localhost/ we get a
503 exception. Please diagnose and solve the issue.
Please deliver:
● A short description of your troubleshooting process.
● Tell us what's wrong
● Explain what you did to fix it